backend/server/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6

# Async task processing
celery==5.3.4
redis==5.0.1
kombu==5.3.4

# AI providers
openai==1.3.7
anthropic==0.7.7
google-genai

# Image processing
Pillow==10.1.0
opencv-python==4.8.1.78

# File handling and validation
aiofiles==23.2.0
python-magic==0.4.27

# Configuration and environment
pydantic[email]==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# HTTP client
httpx==0.25.2
requests==2.31.0
aiohttp==3.9.1

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
flake8==6.1.0

# Monitoring and logging
structlog==23.2.0